PTs Not Working Yet

+++ GSE Control System/Rocket_Side/sensors.h
@@ -10,7 +10,7 @@
 // -------------------------------
 
 // Pressure Transducer Constants
-constexpr uint8_t PT_MaxP[NUM_PT] = {1000,1600,1600};
+constexpr uint32_t PT_MaxP[NUM_PT] = {1000,1600,1600};
 constexpr float PT_MaxV = 4.5; // Maximum reading voltage of pressure transducers
 constexpr float PT_RefV = 5.0; // Used to calculating bin/psi
 constexpr float ADC_Resolution = 1023.0; // Resolution of analog to digital converter
